windows/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic>=2.6.0
streamlit>=1.35.0
plotly>=5.20.0
pandas>=2.1.0
requests>=2.31.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0

scikit-learn>=1.4.0
fpdf2>=2.7.0
networkx>=3.2
